requirements.txt
flask==2.3.3
pytesseract==0.3.10
pillow==10.0.1
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
langdetect==1.0.9
